Move any methods used out of CRUD (i.e. in Convert) to Convert class and change accessibility
