+++ Lusas_Adapter/CRUD/Read/Loads/AreaUniformlyDistributedLoads.cs
@@ -60,7 +60,7 @@ namespace BH.Adapter.Lusas
                     if (lusasDistributedLoad.getValue("type") == "Area")
                     {
                         IEnumerable<IGrouping<string, IFAssignment>> groupedByLoadcases =
-                            LusasAdapter.GetLoadAssignments(lusasDistributedLoad);
+                            GetLoadAssignments(lusasDistributedLoad);
 
                         foreach (IEnumerable<IFAssignment> groupedAssignment in groupedByLoadcases)
                         {
