+++ Lusas_Adapter/CRUD/Read/Loads/BarVaryingDistributedLoads.cs
@@ -52,7 +52,7 @@ namespace BH.Adapter.Lusas
                 {
                     IFLoading lusasInternalBeamDistributedLoad = (IFLoading)lusasInternalBeamDistributedLoads[i];
                     IEnumerable<IGrouping<string, IFAssignment>> groupedByLoadcases =
-                        LusasAdapter.GetLoadAssignments(lusasInternalBeamDistributedLoad);
+                        GetLoadAssignments(lusasInternalBeamDistributedLoad);
 
                     foreach (IEnumerable<IFAssignment> groupedAssignment in groupedByLoadcases)
                     {
