+++ Lusas_Adapter/Lusas_Adapter.csproj
@@ -167,18 +167,18 @@
       <HintPath>..\..\BHoM\Build\Geometry_oM.dll</HintPath>
       <Private>False</Private>
     </Reference>
-	<Reference Include="Lusas.Interop, Version=17.0.1633.29698, Culture=neutral, PublicKeyToken=ee620220e102840a, processorArchitecture=x86" Condition="'$(Configuration)'=='Debug17' Or '$(Configuration)'=='Release17'">
+    <Reference Include="Lusas.Interop, Version=17.0.1633.29698, Culture=neutral, PublicKeyToken=ee620220e102840a, processorArchitecture=x86" Condition="'$(Configuration)'=='Debug17' Or '$(Configuration)'=='Release17'">
       <SpecificVersion>False</SpecificVersion>
       <EmbedInteropTypes>True</EmbedInteropTypes>
       <HintPath>..\libs\Version 17\Lusas.Interop.dll</HintPath>
-		<Private>False</Private>
+      <Private>False</Private>
     </Reference>
     <Reference Include="Lusas.Interop, Version=18.0.1637.31413, Culture=neutral, PublicKeyToken=ee620220e102840a, processorArchitecture=x86" Condition="'$(Configuration)'=='Debug18' Or '$(Configuration)'=='Release18'">
       <SpecificVersion>False</SpecificVersion>
       <EmbedInteropTypes>True</EmbedInteropTypes>
       <HintPath>..\libs\Version 18\Lusas.Interop.dll</HintPath>
       <Private>False</Private>
-    </Reference>	
+    </Reference>
     <Reference Include="Physical_oM">
       <HintPath>..\..\BHoM\Build\Physical_oM.dll</HintPath>
       <Private>False</Private>
@@ -285,16 +285,16 @@
     <Compile Include="CRUD\Read\Results\MeshResults.cs" />
     <Compile Include="CRUD\Read\Results\NodeResults.cs" />
     <Compile Include="CRUD\Read\Results\ReadResults.cs" />
-    <Compile Include="Private Helpers\AssignmentWarning.cs" />
-    <Compile Include="Private Helpers\GetAdapterID.cs" />
-    <Compile Include="Private Helpers\GetAttributeAssignments.cs" />
-    <Compile Include="Private Helpers\GetLineAssignments.cs" />
+    <Compile Include="Convert\ToBHoM\Loads\AssignmentWarning.cs" />
+    <Compile Include="Convert\ToBHoM\GetAdapterID.cs" />
+    <Compile Include="Convert\ToBHoM\Elements\GetAttributeAssignments.cs" />
+    <Compile Include="Convert\ToBHoM\Loads\GetLineAssignments.cs" />
     <Compile Include="Private Helpers\GetLoadAssignments.cs" />
-    <Compile Include="Private Helpers\GetName.cs" />
-    <Compile Include="Private Helpers\GetNode.cs" />
-    <Compile Include="Private Helpers\GetPointAssignments.cs" />
-    <Compile Include="Private Helpers\GetSurfaceAssignments.cs" />
-    <Compile Include="Private Helpers\IsMemberOf.cs" />
+    <Compile Include="Convert\ToBHoM\GetName.cs" />
+    <Compile Include="Convert\ToBHoM\Elements\GetNode.cs" />
+    <Compile Include="Convert\ToBHoM\Loads\GetPointAssignments.cs" />
+    <Compile Include="Convert\ToBHoM\Loads\GetSurfaceAssignments.cs" />
+    <Compile Include="Convert\ToBHoM\Elements\IsMemberOf.cs" />
     <Compile Include="Type\NextFreeId.cs" />
     <Compile Include="CRUD\Delete\Elements\Surface.cs" />
     <Compile Include="Private Helpers\AssignObjectSet.cs" />
